created home component
