+++ src/App.js
@@ -1,25 +1,11 @@
 import './App.css';
-import CockTailItem from './CockTailItem/CockTailItem';
-import { dummy_data } from './DUMMY_DATA';
 import About from "./About/About";
 import NavBar from './NavBar/NavBar'
 import { Route, Switch } from "react-router-dom";
 import Details from './Details/CocktailDetails.js';
+import Home from './Home/Home';
 
 function App() {
-
-
-  let filteredList = dummy_data.map((a) => (
-    <CockTailItem
-      id={a.id}
-      name={a.name}
-      imageName={a.imageName}
-      glass={a.glass}
-      info={a.info}
-    />
-  ));
-
-
   return (
     <>
       <NavBar />
@@ -31,7 +17,7 @@ function App() {
             <Details />
           </Route>
         <Route path="/">
-          {filteredList}
+          <Home />
         </Route>
       </Switch>
     </>
